+++ plugins/gs/src/components/hooks/useHelmReleases.ts
@@ -1,24 +1,7 @@
-import { gsApiRef } from '../../apis';
-import { useApi } from '@backstage/core-plugin-api';
-import { useInstallations } from './useInstallations';
-import { useQueries } from '@tanstack/react-query';
 import { IHelmRelease } from '../../model/services/mapi/helmv2beta1';
-import { getInstallationsQueriesInfo } from './utils';
+import { helmReleaseGVK } from '../../model/services/mapi/objects';
+import { useListResources } from './useListResources';
 
 export function useHelmReleases() {
-  const {
-    selectedInstallations,
-  } = useInstallations();
-  const api = useApi(gsApiRef);
-
-  const queries = useQueries({
-    queries: selectedInstallations.map(installationName => {
-      return {
-        queryKey: [installationName, 'helmreleases'],
-        queryFn: () => api.listHelmReleases({ installationName }),
-      }
-    }),
-  });
-
-  return getInstallationsQueriesInfo<IHelmRelease[]>(selectedInstallations, queries);
+  return useListResources<IHelmRelease>(helmReleaseGVK);
 }