+++ plugins/gs/src/plugin.ts
@@ -1,26 +1,12 @@
 import {
-  configApiRef,
-  createApiFactory,
   createPlugin,
   createRoutableExtension,
-  discoveryApiRef,
 } from '@backstage/core-plugin-api';
 
 import { entityDeploymentsRouteRef, rootRouteRef } from './routes';
-import { GSClient, gsApiRef } from './apis';
-import { scmAuthApiRef } from '@backstage/integration-react';
-import { withQueryClient } from './withQueryClient';
 
 export const gsPlugin = createPlugin({
   id: 'gs',
-  apis: [
-    createApiFactory({
-      api: gsApiRef,
-      deps: { configApi: configApiRef, discoveryApi: discoveryApiRef, scmAuthApi: scmAuthApiRef },
-      factory: ({ configApi, discoveryApi, scmAuthApi }) =>
-        new GSClient({ configApi, discoveryApi, scmAuthApi }),
-    }),
-  ],
   routes: {
     root: rootRouteRef,
     entityContent: entityDeploymentsRouteRef,
@@ -31,7 +17,7 @@ export const GSPluginPage = gsPlugin.provide(
   createRoutableExtension({
     name: 'GSPluginPage',
     component: () =>
-      import('./components/PluginPage').then(m => withQueryClient(m.PluginPage)),
+      import('./components/PluginPage').then(m => m.PluginPage),
     mountPoint: rootRouteRef,
   }),
 );
@@ -39,7 +25,7 @@ export const GSPluginPage = gsPlugin.provide(
 export const EntityGSDeploymentsContent = gsPlugin.provide(
   createRoutableExtension({
     name: 'EntityGSDeploymentsContent',
-    component: () => import('./components/Router').then(m => withQueryClient(m.Router)),
+    component: () => import('./components/Router').then(m => m.Router),
     mountPoint: entityDeploymentsRouteRef,
   }),
 );